+++ src/applications/edu-benefits/0994/components/PaymentReviewView.jsx
@@ -1,6 +1,7 @@
 import _ from 'lodash';
 import React from 'react';
 import { connect } from 'react-redux';
+import { maskBankInformation } from '../utils';
 
 class PaymentReviewView extends React.Component {
   constructor(props) {
@@ -15,6 +16,7 @@ class PaymentReviewView extends React.Component {
     const { formData, name } = this.state;
     let value = formData;
     if (formData === undefined) {
+      // Use prefill data
       const propName = `bank${name
         .substring(0, 1)
         .toUpperCase()}${name.substring(1)}`;
@@ -24,6 +26,10 @@ class PaymentReviewView extends React.Component {
         {},
       );
       value = _.get(prefillBankAccount, propName, '');
+    } else if (name === 'accountNumber' || name === 'routingNumber') {
+      value = maskBankInformation(value, 4);
+    } else if (name === 'accountType' && value.length > 0) {
+      value = value.charAt(0).toUpperCase() + value.substr(1);
     }
     return <span>{value}</span>;
   }
